Adding json serialization to tests
