+++ NitroxTest/NitroxTest.csproj
@@ -97,7 +97,7 @@
     <Compile Include="Patcher\Patches\Equipment_RemoveItem_PatchTest.cs" />
     <Compile Include="Patcher\Test\PatchTestHelper.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Serialization\InteractiveChildObjectIdentifiersSerializationtest.cs" />
+    <Compile Include="Serialization\InteractiveChildObjectIdentifiersSerializationTest.cs" />
     <Compile Include="Serialization\PdaStateDataSerializationTest.cs" />
     <Compile Include="SetupAssemblyInitializer.cs" />
     <Compile Include="Threading\ThreadSafeCollectionTest.cs" />
