+++ NitroxTest/Serialization/PdaStateDataSerializationTest.cs
@@ -27,22 +27,25 @@ namespace NitroxTest.Serialization
         [TestMethod]
         public void Sanity()
         {
-            ServerProtoBufSerializer server = new ServerProtoBufSerializer();
+            IServerSerializer[] serializers = { new ServerProtoBufSerializer(), new ServerJsonSerializer() };
 
-            PDAStateData deserialized;
-            using (MemoryStream stream = new MemoryStream())
+            foreach (IServerSerializer serializer in serializers)
             {
-                server.Serialize(stream, state);
-                stream.Position = 0;
-                deserialized = server.Deserialize<PDAStateData>(stream);
-            }
+                PDAStateData deserialized;
+                using (MemoryStream stream = new MemoryStream())
+                {
+                    serializer.Serialize(stream, state);
+                    stream.Position = 0;
+                    deserialized = serializer.Deserialize<PDAStateData>(stream);
+                }
 
-            deserialized.PdaLog.Count.ShouldBeEquivalentTo(1);
-            deserialized.PdaLog[0].Key.ShouldBeEquivalentTo("Some key");
-            deserialized.EncyclopediaEntries.Count.ShouldBeEquivalentTo(3);
-            deserialized.EncyclopediaEntries[2].ShouldBeEquivalentTo("");
-            deserialized.PartiallyUnlockedByTechType.Count.ShouldBeEquivalentTo(1);
-            deserialized.PartiallyUnlockedByTechType[new NitroxTechType("Battery")].Progress.ShouldBeEquivalentTo(1f);
+                deserialized.PdaLog.Count.ShouldBeEquivalentTo(1);
+                deserialized.PdaLog[0].Key.ShouldBeEquivalentTo("Some key");
+                deserialized.EncyclopediaEntries.Count.ShouldBeEquivalentTo(3);
+                deserialized.EncyclopediaEntries[2].ShouldBeEquivalentTo("");
+                deserialized.PartiallyUnlockedByTechType.Count.ShouldBeEquivalentTo(1);
+                deserialized.PartiallyUnlockedByTechType[new NitroxTechType("Battery")].Progress.ShouldBeEquivalentTo(1f);
+            }
         }
     }
 }
